Don't stop install process if mysqld fails to start (dpkg)
As the final step of dh_installinit, dh tries to start the mysqld server
after installing the init script to /etc/init.d/. If the server fails to
start for any reason (for example a buggy my.cnf), dpkg or apt-get report
a failed installation, which does not reflect what really happened,
creating some confusion.

Fix this by appending --error-handler=: to dh_installinit, effectively
ignoring if the server was able to get started or not -- since this
doesn't really affect the installation process.