Fix - bug 1191580 (InnoDB: Failing assertion:   bitmap_files->files[0].seq_num == first_file_seq_num in file   log0online.cc line 1423 | abort in   log_online_setup_bitmap_file_range); - bug 1191589 (mysqld-debug: .../sql/protocol.cc:518: void   Protocol::end_statement(): Assertion `0' failed on a   INFORMATION_SCHEMA.INNODB_CHANGED_PAGES query).
The issue is bug 1191580 is a race condition between the two bitmap
file enumeration passes in log_online_setup_bitmap_file_range().   If
the 1st bitmap file in the range is deleted at that point, then the
debug build will crash as reported on this bug.  For release builds
this is not an issue as the missing bitmap file will be detected and
diagnosed on an attempt to open it, as fixed by bug 1179974. The
current bug shows that the fix for the latter was incomplete.

Fixed by adjusting the range check code in
log_online_setup_bitmap_file_range() to check for the above condition,
print diagnostics, and return range setup error.  Since these
diagnostics are now printed in two places instead of one, pull it out
to a new static function log_online_diagnose_inconsistent_dir().
Adjust it to print "Warning" instead of "Error" as it's not a fatal
issue.

At the same time fix bug 1191589.  The issue there is that on
log_online_bitmap_iterator_init() failure (such as bug 1191580), the
error is not returned to client properly.  Fixed by returning
ER_CANT_FIND_SYSTEM_REC there and also for a
schema_table_store_record() failure.

At the same time take a related bug 1204075 fix from XtraBackup.

Add a testcase, common for both bugs, to
percona_changed_page_bmp_debug.test.