Implement query thread waits at the sync preflush point, blueprint https://blueprints.launchpad.net/percona-server/+spec/xtradb-preflush-backoff.
Whenever a query thread needs to perform a sync preflush to proceed,
two options are now available:
1) as previously, the query thread may issue a flush list batch itself
and wait for it to complete.  This is also used whenever the page
cleaner thread is not running.
2) alternatively the query thread may wait until the flush list flush
is performed by the page cleaner thread.  The wait is implemented by
maintaining a wait iteration counter, and a random duration sleep of up
to two to power of loop counter value is performed.  To prevent
runaway sleeps, the counter is reset every 16 iterations.

The behavior is controlled by a new system variable
innodb_foreground_preflush with two possible values "sync_preflush"
and "exponential_backoff".  Add a sys_vars test for this variable.