Bug #1200788: InnoDB: Failing assertion: trx_sys->descr_n_used               <= (trx_sys->rw_trx_list).count in file trx0trx.cc line               1492
The problem was in a debug-only assertion introduced when merging the
trx descriptor patch from 5.5.

The assertion’s invariant is supposed to be enforced whenever
trx_sys->rw_trx_list is modified, i.e. when trx_sys->mutex is
locked. Which is already done earlier in the same function. The same
assertion later (with trx_sys->mutex released) is therefore redundant
and has been removed.