Implementation of https://blueprints.launchpad.net/percona-server/+spec/tp-low-prio-queue-throttling-5.5
Introduced a limit on ‘busy’ threads. A busy thread is either in the
active (i.e. executing a statement) or waiting (i.e. between calls to
thd_wait_begin() and thd_wait_end()) state. No events from the low
priority queue are processed when that limit is reached in a thread
group.

Also made the code that creates new threads in thd_wait_begin() a
compile-time option. It’s currently enabled, but benchmarks show that
the code has essentially no effect. That code will likely be removed
later.